+++ app/admin/form_answers.rb
@@ -39,6 +39,17 @@ ActiveAdmin.register FormAnswer do
 
   controller do
     helper :application
+
+    def destroy
+      form_answer = FormAnswer.find(params[:id])
+      if form_answer.signed?
+        flash[:error] = 'Cannot delete signed form answers'
+        redirect_back(fallback_location: admin_form_answers_path)
+        return
+      end
+
+      destroy!
+    end
   end
 
   index do